Redesigned core structure
Change the whole core structure. From using tasks only i started using pre_tasks and post_tasks effectively. Created more tasks that can automatically create directory to make the script with minimum errors. After running the script the playbook will automatically clear files and created directories that is created. If the user choosen directory with files already this directory will not be deleted due to potential data loss. This means that there is a mechanism that checks if there is files in the working directory or not, if there are files, the other files in the directories recursively will not be removed. This features was tested many times and i did not see any data loss or curruption. But still. Use will use this playbook at your own risk!